+++ CHANGELOG.md
@@ -1,2 +1,1 @@
-- Shifted code from `MainActivity.kt` to `LoginScreen.kt`
-- Optimized imports
+- Built `ProfileScreen.kt`